+++ genium/src/test/resources/smoke/defaults/output/android/com/example/smoke/StructWithAnEnum.java
@@ -1,0 +1,16 @@
+/*
+ *
+ * Automatically generated. Do not modify. Your changes will be lost.
+ */
+package com.example.smoke;
+import android.support.annotation.NonNull;
+public class StructWithAnEnum {
+    @NonNull
+    public AnEnum config;
+    public StructWithAnEnum() {
+        this.config = AnEnum.ENABLED;
+    }
+    public StructWithAnEnum(@NonNull final AnEnum config) {
+        this.config = config;
+    }
+}